backend/requirements.txt
altgraph==0.17.2
androguard==3.4.0a1
annotated-types==0.6.0
apkid==2.1.3
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.0.8
attrs==23.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bandit==1.7.8
beautifulsoup4==4.11.1
biplist==1.0.3
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cli-helpers==2.2.1
click==8.0.1
cmake==3.27.9
colorama==0.4.5
colorlog==6.7.0
configobj==5.0.6
cryptography==41.0.7
cycler==0.11.0
decorator==4.4.2
delegator.py==0.1.1
distlib==0.3.8
distro==1.7.0
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jsonapi==7.0.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
exceptiongroup==1.2.1
executing==1.0.0
filelock==3.14.0
Flask==2.0.3
fonttools==4.37.2
frida==15.2.2
frida-tools==11.0.0
google-play-scraper==1.2.2
graphviz==0.20.1
h11==0.13.0
h2==4.1.0
hpack==4.0.0
http-tools==2.1.0
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
IP2Location==8.8.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kaitaistruct==0.9
kaleido==0.2.1
kiwisolver==1.4.4
ldap3==2.9.1
libsast==1.5.2
lief==0.12.1
litecli==1.9.0
lxml==4.9.1
macholib==1.16
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mitmproxy==8.0.0
msgpack==1.0.4
nassl==5.1.0
networkx==2.8.6
numpy==1.23.3
objection==1.11.0
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==21.3
pandas==1.4.4
parso==0.8.3
passlib==1.7.4
pbr==6.0.0
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2023.12.1
platformdirs==4.2.2
plotly==5.10.0
prettytable==3.4.1
prompt-toolkit==3.0.19
protobuf==3.19.5
psutil==5.9.2
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.3.0
pydantic_core==2.6.3
pydivert==2.1.0
pydot==1.4.2
Pygments==2.13.0
PyJWT==2.8.0
pymongo==3.12.1
pyOpenSSL==23.3.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
quark-engine==22.6.1
requests==2.28.1
requests-file==1.5.1
rich==13.7.1
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
rzpipe==0.4.0
selenium==4.20.0
semver==2.13.0
shelljob==0.6.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
sslyze==5.2.0
stack-data==0.5.0
stevedore==5.2.0
tabulate==0.8.10
tenacity==8.0.1
tldextract==3.3.1
tls-parser==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2022.2
urllib3==1.26.12
urwid==2.1.2
virtualenv==20.26.2
waitress==2.1.2
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==6.2.0
wsproto==1.1.0
yara-python-dex==1.0.3
zstandard==0.17.0
